+++ windows/vk_native_client_plugin.h
@@ -31,6 +31,9 @@ namespace vk_native_client
         void SetHTMLClipboard(
             const flutter::MethodCall<flutter::EncodableValue> &method_call,
             std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
+        void canCopyFromClipboard(
+            const flutter::MethodCall<flutter::EncodableValue> &method_call,
+            std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
     };
 
 } // namespace vk_native_client